move checkSSLAllowed to client implemantation
